Replaced cy.log() with nested describe/it
The test-from-scenario-step generator was changed recently to use cy.log() instead of describe/it.
This made displaying test scenarios on cy commands or console to not aggregate results under their steps.
Rather it was adding all the results on the same level under scenario title with printouts of the Given when then steps.