fix(core): fix usage of configFile param
Because of a system behavior in linux it was not possible for the preprocessor to determine if
cypress were started with a different config file location. The reason was linux changing the order
of command line arguments which separated --config-file from the actual value. To fix this I let the
user pass the cypress config during initialization phase and therefore prevent the need to read the
config.json at all.

fix #283