feat(cypress-tags.js): provide cypress-executable through cypress.json
As sometimes it might be necessary to provide a wrapper for cypress (see https://github.com/sorry-cypress/cy2) there should also be a possibility to divert from the fixed usage of the cypress-executable. To do so an easy approach would be to check for the entry `cypressExecutable` within the cypress.json. If this is set, it is used instead the default cypress-executable.