fix(cypress-tags.js): support for arbitrary glob patterns in `testFiles` in `cypress.json`
cypress-tags finds no files when `testFiles` is set to `**/*.{feature,features}` in `cypress.json`.
The problem was that the script expected that we can specify multiple glob patterns separated by a
comma. But that is wrong. Cypress requires us to specify either an array of glob patterns or a
single glob pattern string. And those glob patterns can be as complex as we want. So when a single
string is specified, splitting by `,` was just wrong.

Tested patterns:

- `"testFiles": "**/*.feature"`
- `"testFiles": "**/*.{feature,features}"`
- `"testFiles": "{**/*.{features,featurex},**/*.feature}"`
- `"testFiles": ["**/*.{feature,features}", "**/*.feature"]`
- `"testFiles": ["{**/*.{features,featurex},**/*.feature}", "**/*.feature"]`

This fix is related to #196.